# packages/backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the API server for the Diplom project",
        "title": "Diplom Backend API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8082",
    "basePath": "/api",
    "paths": {
...
        "/publications": {
            "get": {
                "description": "Get a list of all publications",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "publications"
                ],
                "summary": "Get all publications",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Publication"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new publication",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "publications"
                ],
                "summary": "Create a new publication",
                "parameters": [
                    {
                        "description": "Publication object",
                        "name": "publication",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Publication"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Publication"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/publications/{id}": {
            "get": {
                "description": "Get a single publication by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "publications"
                ],
                "summary": "Get a publication by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Publication ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Publication"
                        }
                    },
                    "400": {
                        "description": "Invalid publication ID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Publication not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "description": "Update an existing publication's information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "publications"
                ],
                "summary": "Update a publication",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Publication ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Publication object",
                        "name": "publication",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Publication"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Publication"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a publication by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "publications"
                ],
                "summary": "Delete a publication",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Publication ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/publications/{id}/authors": {
            "get": {
                "description": "Get all researchers who are authors of a specific publication",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "publications"
                ],
                "summary": "Get authors of a publication",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Publication ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Researcher"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid publication ID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Publication not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.LocalizedString": {
            "type": "object",
            "properties": {
                "en": {
                    "type": "string"
                },
                "ru": {
                    "type": "string"
                }
            }
        },
...
        "models.Publication": {
            "type": "object",
            "properties": {
                "authors": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "journal": {
                    "type": "string"
                },
                "link": {
                    "type": "string"
                },
                "title": {
                    "$ref": "#/definitions/models.LocalizedString"
                },
                "year": {
                    "type": "integer"
                }
            }
        },
        "models.Researcher": {
            "type": "object",
            "properties": {
                "bio": {
                    "$ref": "#/definitions/models.LocalizedString"
                },
                "id": {
                    "type": "integer"
                },
                "lastName": {
                    "$ref": "#/definitions/models.LocalizedString"
                },
                "name": {
                    "$ref": "#/definitions/models.LocalizedString"
                },
                "photo": {
                    "type": "string"
                },
                "profiles": {
                    "$ref": "#/definitions/models.ResearcherProfiles"
                },
                "publications": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Publication"
                    }
                }
            }
        },
        "models.ResearcherProfiles": {
            "type": "object",
            "properties": {
                "googleScholar": {
                    "type": "string"
                },
                "orcid": {
                    "type": "string"
                },
                "publons": {
                    "type": "string"
                },
                "researchgate": {
                    "type": "string"
                },
                "scopus": {
                    "type": "string"
                }
            }
        },
...
    }
}